+++ mc/CodeAnalysis/Binding/BoundBinaryExpression.cs
@@ -1,0 +1,18 @@
+namespace Minsk.CodeAnalysis.Binding
+{
+    internal sealed class BoundBinaryExpression : BoundExpression
+    {
+        public BoundBinaryExpression(BoundExpression left, BoundBinaryOperatorKind operatorKind, BoundExpression right)
+        {
+            Left = left;
+            OperatorKind = operatorKind;
+            Right = right;
+        }
+        public override BoundNodeKind Kind => BoundNodeKind.LiteralExpression;
+        public override Type Type => Left.Type;
+
+        public BoundExpression Left { get; }
+        public BoundBinaryOperatorKind OperatorKind { get; }
+        public BoundExpression Right { get; }
+    }
+}
