+++ mc/CodeAnalysis/Binding/BoundBinaryOperatorKind.cs
@@ -1,0 +1,11 @@
+namespace Minsk.CodeAnalysis.Binding
+{
+    internal enum BoundBinaryOperatorKind
+    {
+        Addition,
+        Subtraction,
+        Multiplication,
+        Division,
+
+    }
+}
