+++ mc/CodeAnalysis/Binding/BoundExpression.cs
@@ -1,0 +1,7 @@
+namespace Minsk.CodeAnalysis.Binding
+{
+    internal abstract class BoundExpression : BoundNode
+    {
+        public abstract Type Type { get; }
+    }
+}
