+++ mc/CodeAnalysis/Binding/BoundNodeKind.cs
@@ -1,0 +1,9 @@
+namespace Minsk.CodeAnalysis.Binding
+{
+
+    internal enum BoundNodeKind
+    {
+        LiteralExpression,
+        UnaryExpression
+    }
+}
