+++ mc/CodeAnalysis/Binding/BoundUnaryExpression.cs
@@ -1,0 +1,17 @@
+namespace Minsk.CodeAnalysis.Binding
+{
+    internal sealed class BoundUnaryExpression : BoundExpression
+    {
+
+        public BoundUnaryExpression(BoundUnaryOperatorKind operatorKind, BoundExpression operand)
+        {
+            OperatorKind = operatorKind;
+            Operand = operand;
+        }
+
+        public override BoundNodeKind Kind => BoundNodeKind.UnaryExpression;
+        public override Type Type => Operand.Type;
+        public BoundUnaryOperatorKind OperatorKind { get; }
+        public BoundExpression Operand { get; }
+    }
+}
